+++ src/defaults.ts
@@ -9,8 +9,8 @@ export const host = process.env.PGHOST || 'localhost';
 export const port = parseInt(process.env.PGPORT as string, 10) || 5432;
 export const user = process.env.PGUSER || (
     process.platform === 'win32' ? process.env.USERNAME : process.env.USER
-);
-export const database = process.env.PGDATABASE || 'postgres';
+) as string;
+export const database = process.env.PGDATABASE;
 export const password = process.env.PGPASSWORD;
 export const preparedStatementPrefix = 'tsp_';
 export const sslMode = process.env.PGSSLMODE;